Write crash logs to non-volatile storage
This aids in debugging failures that cause the device to become
unreachable over the network by reading the data partition using a
different device. Regular logs cause dangerously many writes to the SD
cards of Raspberry Pis and will remain in volatile storage. If
necessary, services should switch to eprintln for non-volatile logging.